requirements.txt
python-dotenv==1.0.1
python-telegram-bot==20.8
Requests==2.31.0
